Separate components into each file & Improve code readability
