+++ src/components/VoteButton/VoteButton.js
@@ -1,10 +1,9 @@
-import React, { Component, memo } from 'react'
+import React, { Component } from 'react'
 import { isEmpty } from 'lodash'
 import { withRouter } from 'react-router'
 import PropTypes from 'prop-types'
-import CircularProgress from '@material-ui/core/CircularProgress'
-import { Grid, Grow, Typography, Portal, Tooltip, SvgIcon, Snackbar } from '@material-ui/core'
-import { withStyles, useTheme, withTheme } from '@material-ui/core/styles'
+import { Grid, Grow, Portal, Tooltip, SvgIcon, Snackbar } from '@material-ui/core'
+import { withStyles } from '@material-ui/core/styles'
 import SnackbarContent from '@material-ui/core/SnackbarContent'
 import polly from 'polly-js'
 import numeral from 'numeral'
@@ -13,527 +12,28 @@ import axios from 'axios'
 import { parseError } from '../../eos/error'
 import { connect } from 'react-redux'
 import { setPostInfo, updateInitialVote, updateVoteLoading } from '../../redux/actions'
-import { levelColors } from '../../utils/colors'
-import Rating from '@material-ui/lab/Rating'
-import equal from 'fast-deep-equal'
 import WelcomeDialog from '../WelcomeDialog/WelcomeDialog'
-import ErrorBoundary from '../ErrorBoundary/ErrorBoundary'
 import scatter from '../../eos/scatter/scatter.wallet'
 import rollbar from '../../utils/rollbar'
 import isEqual from 'lodash/isEqual'
 import { accountInfoSelector, ethAuthSelector } from '../../redux/selectors'
 import { deletevote, editvote, createvotev4, postvotev4, postvotev3, createvote } from '../../eos/actions/vote'
+import PostStats from './PostStats'
+import IconContainer from './IconContainer'
+import {
+  CAT_DESC,
+  CREATE_VOTE_LIMIT,
+  DEFAULT_WAIT_AND_RETRY,
+  dislikeRatingConversion,
+  likeRatingConversion,
+  ratingConversion
+} from './constants'
+import { convertRating } from './helpers'
+import StyledRating from './StyledRating'
+import CatIcon from './CatIcon'
+import styles from './styles'
 
 const { BACKEND_API } = process.env
-const ICONS = process.env.ICONS.split(',')
-const CREATE_VOTE_LIMIT = 20
-
-const CAT_ICONS = {
-  popularity: ICONS[0],
-  intelligence: ICONS[1],
-  funny: ICONS[2],
-  useful: ICONS[3],
-  knowledgeable: ICONS[4],
-  interesting: ICONS[5],
-  expensive: ICONS[6],
-  engaging: ICONS[7],
-  easy: ICONS[8],
-  chill: ICONS[9],
-  beautiful: ICONS[10],
-  affordable: ICONS[11],
-  trustworthy: ICONS[12],
-  wouldelect: ICONS[13],
-  agreewith: ICONS[14],
-  original: ICONS[15],
-  fire: ICONS[16]
-}
-
-const CAT_DESC = {
-  easy:
-    'Easy: can do well without extra effort; generous grading, minimal time',
-  interesting: 'Interesting: compelling subject matter, makes you think',
-  useful: 'Useful: has important knowledge for your field/career',
-  knowledgeable:
-    "Knowledgeable: knows what they're talking about, expert in subject",
-  engaging:
-    'Engaging: captures your attention, makes concepts easy to understand',
-  chill: 'Chill: cool, laidback, a vibe',
-  popularity: 'like',
-  intelligence: 'smart',
-  trustworthy: 'most trustworthy',
-  wouldelect: 'most electable',
-  agreewith: 'most agreed with',
-  fire: 'Fire: really good, amazing'
-}
-
-const DEFAULT_WAIT_AND_RETRY = [
-  250,
-  250,
-  250,
-  250,
-  250,
-  300,
-  350,
-  400,
-  400,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500,
-  500
-]
-
-const styles = (theme) => ({
-  greenArrow: {
-    color: levelColors.second
-  },
-  redArrow: {
-    color: levelColors.sixth
-  },
-  defaultArrow: {
-    color: 'white',
-    opacity: 0.6
-  },
-  catIcon: {
-    width: 20,
-    height: 20,
-    borderRadius: '50%',
-    padding: '2px',
-    [theme.breakpoints.down('xs')]: {
-      height: 25,
-      width: 25,
-      margin: '0'
-    }
-  },
-  postWeight: {
-    minWidth: '50px',
-    fontSize: '16px',
-    [theme.breakpoints.down('xs')]: {
-      fontSize: '20px'
-    }
-  },
-  snackbar: {
-    position: 'absolute',
-    backgroundColor: '#ff5252',
-    textColor: '#f0f0f0',
-    width: '8%'
-  },
-  snack: {
-    backgroundColor: '#ff5252',
-    color: '#fff8f3',
-    fontWeight: 'light'
-  },
-  snackbarContent: {
-    width: 150
-  },
-  snackUpper: {
-    backgroundColor: 'transparent',
-    paddingBottom: 0
-  },
-  dialog: {
-    width: '100%',
-    marginLeft: 190,
-    [theme.breakpoints.down('md')]: {
-      marginLeft: 0,
-      width: '100%'
-    },
-    [theme.breakpoints.up('md')]: {
-      marginLeft: 190,
-      width: `calc(100% - 190px)`
-    },
-    [theme.breakpoints.up('1600')]: {
-      width: '100%',
-      marginLeft: 0
-    }
-  },
-  mobileBtn: {
-    [theme.breakpoints.down('sm')]: {
-      width: '1.2em'
-    }
-  }
-})
-
-const ratingStyles = ({ palette }) => ({
-  iconFilled: {
-    border: '3px',
-    borderColor: palette.common.first,
-    color: palette.alt.first
-  },
-  iconHover: {
-    stroke: palette.common.first
-  },
-  iconEmpty: {
-    color: palette.alt.first
-  }
-})
-const StyledRating = withStyles(ratingStyles)(Rating)
-
-const labels = {
-  1: '1',
-  2: '2',
-  3: '3',
-  4: '4',
-  5: '5'
-}
-
-const ratingConversion = {
-  1: 2,
-  2: 1,
-  3: 1,
-  4: 2,
-  5: 3
-}
-
-const quantileToRating = {
-  first: 5,
-  second: 4,
-  third: 3,
-  fourth: 2,
-  fifth: 1
-}
-
-const ratingToQuantile = {
-  5: 'first',
-  4: 'second',
-  3: 'third',
-  2: 'fourth',
-  1: 'fifth'
-}
-
-const dislikeRatingConversion = {
-  1: 2,
-  2: 1
-}
-
-const likeRatingConversion = {
-  1: 3,
-  2: 4,
-  3: 5
-}
-
-const convertRating = (like, rating) =>
-  like ? likeRatingConversion[rating] : dislikeRatingConversion[rating]
-
-const IconWithRef = class IconWithRef extends Component {
-  iconRef = React.createRef();
-
-  shouldComponentUpdate (nextProps, nextState) {
-    if (!isEqual(nextProps.value, this.props.value) || !isEqual(nextProps.style, this.props.style) || !isEqual(nextState, this.state)) {
-      return true
-    }
-    return false
-  }
-
-  render () {
-    const { value, handleRatingChange } = this.props
-
-    return (
-      <ErrorBoundary>
-        <div
-          ref={this.iconRef} // Refs and props for tooltip + vote mouse events
-          onTouchStart={(e) => {
-            handleRatingChange(e, value)
-          }}
-          onClick={(e) => {
-            handleRatingChange(e, value)
-          }}
-        >
-          <div {...this.props} />
-        </div>
-      </ErrorBoundary>
-    )
-    }
-  }
-
-IconWithRef.propTypes = {
-  handleRatingChange: PropTypes.func.isRequired,
-  value: PropTypes.number.isRequired,
-  style: PropTypes.object.isRequired
-}
-
-const IconContainer = memo((props) => {
-  const { value, ratingAvg, quantile, vote, handleRatingChange, hoverValue } = props
-  const { palette } = useTheme()
-  const quantileColor = levelColors[quantile]
-  const convertedVoterRating = vote
-    ? convertRating(vote.like, vote.rating)
-    : null
-
-  const ratingQuantile = quantileToRating[quantile]
-  const ratingQuantileStyle = (ratingQuantile >= value) ? { color: quantileColor } : { color: palette.alt.third }
-  const voteStyle = (convertedVoterRating >= value) ? { stroke: palette.common.third } : {}
-  const marginStyle = (window.innerWidth <= 440)
-      ? { marginTop: '-3px', marginRight: '-5px', marginLeft: '-6px' }
-      : { marginTop: '-3px', marginRight: '-9px', marginLeft: '-1.5px' }
-
-  const defaultQuantileColor = levelColors[ratingToQuantile[hoverValue]]
-  const hoverStyle = (defaultQuantileColor && hoverValue && hoverValue) >= value
-      ? { color: defaultQuantileColor }
-      : {}
-
-  const style = {
-    ...marginStyle,
-    ...ratingQuantileStyle,
-    ...voteStyle,
-    ...hoverStyle // will override the ratingQuantileStyle if defined
-  }
-
-  return (
-    <StyledTooltip title={labels[value] || ''}
-      enterDelay={1500}
-    >
-      <IconWithRef
-        {...props}
-        value={value}
-        ratingAvg={ratingAvg}
-        quantile={quantile}
-        handleRatingChange={handleRatingChange}
-        style={style}
-      />
-    </StyledTooltip>
-  )
-})
-
-const StyledTooltip = memo(
-  withStyles({
-    popper: {
-      marginTop: '-10px',
-      marginLeft: '14px'
-    }
-  })((props) => (
-    <Tooltip
-      {...props}
-      disableTouchListener
-      classes={{
-        popper: props.classes.popper
-      }}
-    />
-  ))
-)
-
-StyledTooltip.propTypes = {
-  classes: PropTypes.object.isRequired
-}
-
-IconContainer.propTypes = {
-  classes: PropTypes.object.isRequired,
-  value: PropTypes.number.isRequired,
-  handleRatingChange: PropTypes.func.isRequired,
-  ratingAvg: PropTypes.number.isRequired,
-  hoverValue: PropTypes.number.isRequired,
-  quantile: PropTypes.string.isRequired,
-  vote: PropTypes.object
-}
-
-const VoteLoader = (props) => (
-  <CircularProgress size={30}
-    style={{ marginRight: '5px' }}
-  />
-)
-
-class CatIcon extends Component {
-  state = {
-    category: this.props.category,
-    voteLoading: this.props.voteLoading
-  }
-
-  componentDidUpdate (prevProps) {
-    const { quantile, category, voteLoading } = this.props
-    const {
-      quantile: prevQuantile,
-      category: prevCategory,
-      voteLoading: prevVoteLoading
-    } = prevProps
-    if (
-      !equal(
-        {
-          quantile,
-          category,
-          voteLoading
-        },
-        {
-          quantile: prevQuantile,
-          category: prevCategory,
-          voteLoading: prevVoteLoading
-        }
-      )
-    ) {
-      this.updateIconInfo({
-        quantile,
-        category,
-        voteLoading
-      })
-    }
-  }
-
-  updateIconInfo ({ quantile, category, voteLoading }) {
-    this.setState({ category, voteLoading })
-  }
-
-  render () {
-    const { category, voteLoading } = this.state
-    const { classes, handleDefaultVote } = this.props
-
-    if (voteLoading) {
-      return <VoteLoader />
-    }
-
-    return (
-      <Typography className={classes.catIcon}
-        onClick={handleDefaultVote}
-        variant='h4'
-        style={{ fontSize: window.innerWidth <= 600 ? '16px' : 'inherit' }}
-      >
-        {CAT_ICONS[category]}
-      </Typography>
-    )
-  }
-}
-
-CatIcon.propTypes = {
-  quantile: PropTypes.string.isRequired,
-  category: PropTypes.string.isRequired,
-  classes: PropTypes.object.isRequired,
-  handleDefaultVote: PropTypes.func.isRequired,
-  voteLoading: PropTypes.bool.isRequired
-}
-
-const StyledCatIcon = withStyles({
-  catIcon: {
-    width: 35,
-    height: 35,
-    margin: 0,
-    marginTop: '0px',
-    cursor: 'pointer'
-  }
-})(CatIcon)
-
-class PostStats extends Component {
-  state = {
-    weight: this.props.weight,
-    totalVoters: this.props.totalVoters
-  };
-
-  componentDidUpdate (prevProps) {
-    const { weight, totalVoters } = this.props
-    const { weight: prevWeight, totalVoters: prevTotalVoters } = prevProps
-    if (
-      !equal(
-        {
-          weight,
-          totalVoters
-        },
-        {
-          weight: prevWeight,
-          totalVoters: prevTotalVoters
-        }
-      )
-    ) {
-      this.updatePostStats({
-        weight,
-        totalVoters
-      })
-    }
-  }
-
-  updatePostStats ({ weight, totalVoters }) {
-    this.setState({ weight, totalVoters })
-  }
-
-  render () {
-    const { classes, isShown, quantile, theme } = this.props
-    const { totalVoters, weight } = this.state
-    return (
-      <Grid container
-        spacing={0}
-      >
-        <Grid item>
-          <Tooltip title='Post Yup Score'
-            disableTouchListener
-          >
-            <p className={classes.weight}
-              style={{ color: !isShown ? levelColors[quantile] : theme.palette.common.first }}
-            >{weight}</p>
-          </Tooltip>
-        </Grid>
-        <Grid item>
-          <Tooltip title='Number of Voters'
-            disableTouchListener
-          >
-            <p className={classes.totalVoters}>
-              {totalVoters}
-            </p>
-          </Tooltip>
-        </Grid>
-      </Grid>
-    )
-  }
-}
-
-PostStats.propTypes = {
-  classes: PropTypes.object.isRequired,
-  theme: PropTypes.object.isRequired,
-  totalVoters: PropTypes.number.isRequired,
-  weight: PropTypes.number.isRequired,
-  isShown: PropTypes.bool.isRequired,
-  quantile: PropTypes.string.isRequired
-}
-
-const postStatStyles = theme => ({
-  weight: {
-    marginRight: '3px',
-    fontSize: '16px'
-  },
-  totalVoters: {
-    fontSize: '16px',
-    color: theme.palette.common.third,
-    opacity: 0.3,
-    marginLeft: '7px'
-  }
-})
-
-const StyledPostStats = withTheme(withStyles(postStatStyles)(PostStats))
 
 class VoteButton extends Component {
   state = {
@@ -1036,7 +536,7 @@ class VoteButton extends Component {
                   justify='space-around'
                 >
                   <Grid item>
-                    <StyledCatIcon
+                    <CatIcon
                       category={category}
                       handleDefaultVote={this.handleDefaultVote}
                       voteLoading={voteLoading}
@@ -1119,7 +619,7 @@ class VoteButton extends Component {
                           marginRight: '12px'
                         }}
                     >
-                      <StyledPostStats
+                      <PostStats
                         style={{ marginLeft: '15px' }}
                         totalVoters={currTotalVoters}
                         weight={formattedWeight}
