+++ src/components/VoteButton/helpers.js
@@ -1,0 +1,4 @@
+import { dislikeRatingConversion, likeRatingConversion } from './constants'
+
+export const convertRating = (like, rating) =>
+  like ? likeRatingConversion[rating] : dislikeRatingConversion[rating]
